+++ trixnity-messenger/src/androidMain/kotlin/de/connect2x/trixnity/messenger/util/Images.android.kt
@@ -1,10 +1,11 @@
 package de.connect2x.trixnity.messenger.util
 
 import android.graphics.BitmapFactory
-import de.connect2x.trixnity.messenger.viewmodel.util.limitedByteArrayOrNull
 import net.folivo.trixnity.utils.ByteArrayFlow
+import net.folivo.trixnity.utils.toByteArray
 import org.koin.core.module.Module
 import org.koin.dsl.module
+import net.folivo.trixnity.utils.toByteArray
 
 actual fun platformGetImageDimensionsModule(): Module = module {
     single<GetImageDimensions> {
@@ -16,7 +17,7 @@ actual fun platformGetImageDimensionsModule(): Module = module {
 
 
 suspend fun getImageDimensions(byteArrayFlow: ByteArrayFlow, maxMediaSize: Long): Pair<Int?, Int?> {
-    val byteArray = byteArrayFlow.limitedByteArrayOrNull(maxMediaSize)
+    val byteArray = byteArrayFlow.toByteArray(maxMediaSize)
     val bitmap = byteArray?.let { BitmapFactory.decodeByteArray(it, 0, it.size) }
     return bitmap?.width to bitmap?.height
 }