+++ trixnity-messenger/src/commonMain/kotlin/de/connect2x/trixnity/messenger/DI.kt
@@ -200,7 +200,6 @@ fun createTrixnityMessengerDefaultModuleFactories(): List<ModuleFactory> = listO
                     markOwnMessageAsRead = true
                     httpClientEngine = config.httpClientEngine
                     httpClientConfig = config.httpClientConfig
-                    deleteRoomsOnLeave = false
                     lastRelevantEventFilter =
                         { relevantTimelineEvents.isRelevantTimelineEvent(it.content) }
                     if (eventContentSerializerMappings.isNotEmpty()) {
@@ -245,7 +244,7 @@ fun createTrixnityMessengerDefaultModuleFactories(): List<ModuleFactory> = listO
             single<DirectRoom> { DirectRoomImpl() }
             single<ActiveVerifications> { ActiveVerificationsImpl() }
             single<RoomPresence> { RoomPresenceImpl(get()) }
-            single<Search> { SearchImpl(get(), get()) }
+            single<Search> { SearchImpl(get(), get(), get()) }
             single<RunInitialSync> { RunInitialSync }
             single<DragAndDropHandler> { DragAndDropHandlerBase() }
             single<AccountSetupViewModelFactory> { AccountSetupViewModelFactory }
@@ -285,7 +284,7 @@ fun createTrixnityMessengerDefaultModuleFactories(): List<ModuleFactory> = listO
     ::platformPathsModule,
     ::platformStringsModule,
     ::platformCreateRepositoriesModuleModule,
-    ::platformCreateMediaStoreModule,
+    ::platformCreateMediaStoreModuleModule,
     ::secretsModule,
     ::platformGetSystemLangModule,
     ::platformDeleteAccountDataModule,